+++ app/.babelrc
@@ -1,0 +1,4 @@
+{
+    "presets": ["es2015", "react"],
+    "plugins": ["transform-object-rest-spread"]
+}
