+++ app/constants/routes.js
@@ -1,4 +1,4 @@
 // react-router routes
 export default {
-	HOME: '/'
+    HOME: '/'
 }
