+++ app/pages/index.js
@@ -2,6 +2,6 @@ import Navigator from './navigator.jsx';
 import HomePage from './home.jsx';
 
 export {
-	Navigator,
-	HomePage
+    Navigator,
+    HomePage
 }
