+++ app/pages/navigator.jsx
@@ -4,19 +4,19 @@ import { ThemeProvider } from 'styled-components'
 
 class Navigator extends React.Component {
 
-	render () {
-		return (
-			<ThemeProvider theme={this.props.theme}>
-				{this.props.children}
-			</ThemeProvider>
-		)
-	}
+    render () {
+        return (
+            <ThemeProvider theme={this.props.theme}>
+                {this.props.children}
+            </ThemeProvider>
+        )
+    }
 }
 
 function select(state){
-	return {
-		theme: state.theme.data
-	}
+    return {
+        theme: state.theme.data
+    }
 }
 
 export default connect(select)(Navigator)
