+++ app/reducers/index.js
@@ -4,8 +4,8 @@ import { routerReducer } from 'react-router-redux'
 import { theme } from './theme_reducer.js'
 
 const rootReducer = combineReducers({
-	router: routerReducer,
-	theme
+    router: routerReducer,
+    theme
 })
 
 export default rootReducer
