+++ app/reducers/initial_states.js
@@ -1,0 +1,25 @@
+const defaultState = {
+    fetching: false,
+    fetched: false,
+    error: null
+}
+
+const arrayState = {
+    fetching: false,
+    fetched: false,
+    error: null,
+    data: []
+}
+
+const objectState = {
+    fetching: false,
+    fetched: false,
+    error: null,
+    data: {}
+}
+
+export {
+    defaultState,
+    arrayState,
+    objectState
+}
