--- public/.babelrc
@@ -1,4 +1,0 @@
-{
-	"presets": ["es2015", "react"],
-	"plugins": ["transform-object-rest-spread"]
-}
