--- public/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta http-equiv="Content-type" content="text/html; charset=utf-8"/>
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	
-    <title>MHacks 10</title>
-</head>
-<body>
-	<div id="app"></div>
-</body>
-</html>
