--- public/styles/theme.js
@@ -1,12 +1,0 @@
-// example use
-export default {
-	primary1: 'blue',
-	primary2: '',
-	primary3: '',
-	secondary1: '',
-	secondary2: '',
-	secondary3: '',
-	background1: '',
-	background2: '',
-	background3: ''
-}
